Refactor code, WIP
